+++ packages/std/src/storage_keys/mod.rs
@@ -1,5 +1,9 @@
 mod length_prefixed;
+#[cfg(all(feature = "cosmwasm_3_0", feature = "iterator"))]
+mod range;
 
 // Please note that the entire storage_keys module is public. So be careful
 // when adding elements here.
 pub use length_prefixed::{namespace_with_key, to_length_prefixed, to_length_prefixed_nested};
+#[cfg(all(feature = "cosmwasm_3_0", feature = "iterator"))]
+pub(crate) use range::{range_to_bounds, ToByteVec};